requirements.txt
Flask
psycopg2-bin
Jinja2
markupsafe